Rename watermark runtime to manager
